+++ component-test/src/test/kotlin/com/revolut/wallet/core/Transfer.kt
@@ -1,18 +1,17 @@
 package com.revolut.wallet.core
 
 import com.fasterxml.jackson.annotation.JsonInclude
-import java.math.BigDecimal
 import java.util.UUID
 
 @JsonInclude(JsonInclude.Include.NON_EMPTY)
 data class Transfer(
     val fromId: UUID,
     val toId: UUID,
-    val amount: BigDecimal
+    val amount: Double
 ) {
     constructor(
         fromAccount: Account,
         toAccount: Account,
-        amount: BigDecimal
-    ) : this(fromAccount.id, toAccount.id, amount)
+        amount: Long
+    ) : this(fromAccount.id, toAccount.id, amount.toDouble())
 }
